Refactor universal checkpointing and tensor fragments
