+++ deepspeed/utils/__init__.py
@@ -4,4 +4,6 @@ from .comms_logging import get_caller_func
 from .init_on_device import OnDevice
 from .groups import *
 from .nvtx import instrument_w_nvtx
+from .tensor_fragment import tensor_fragment, get_full_hp_param, get_hp_fragment_mapping
+from .mixed_precision_linkage import link_hp_params
 from deepspeed.runtime.dataloader import RepeatingLoader
